% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_choice.R
\name{check_choice}
\alias{check_choice}
\title{Checks if the choice is within the length of a character vector}
\usage{
check_choice(choices, choice)
}
\arguments{
\item{choices}{character vector.}

\item{choice}{numeric or character.  Character values will convert to 0 in the function.}
}
\value{
If the choice is a number within the length range of the choices the return value will be TRUE.
If the choice is a number but is not within the length range of the choice the return value will be FALSE.
If the choices argument is not a character vector with a length of at least one then the program will stop.
}
\description{
Helper function for console_menu.
}
\examples{

\dontrun{
check_choice(c("a", "b"), 1)

check_choice(c("a", "b"), 2)

check_choice(c("a", "b"), 3)

check_choice(c("a", "b"), "a")
}
}
