% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_file_menu.R
\name{console_file_menu}
\alias{console_file_menu}
\title{Prints out a menu from a list of file names in a folder for user selection
wrapper function to console_menu}
\usage{
console_file_menu(
  folder_path = ".",
  pattern = NULL,
 
    instruction_msg = "Please type in the console your choice from above and press enter: ",
  quit_key = "q",
  quit_message = paste0("To quit please type ", quit_key, " and press return"),
  return_number = FALSE,
  test_args = list(test_mode = FALSE, test_input = NA_character_)
)
}
\arguments{
\item{folder_path}{character}

\item{pattern}{character. Regex pattern used in list.files}

\item{instruction_msg}{character.  Message shows under the menu selection}

\item{quit_key}{character.  Character value for the user to quit the menu}

\item{quit_message}{character.  Character value to explain how to quit the menu}

\item{return_number}{logical.  TRUE return number choice.  FALSE returns the index of chr_vector}

\item{test_args}{list(test_mode logical, test_input character).  Only use for testing.}
}
\value{
character
}
\description{
Prints out a menu from a list of file names in a folder for user selection
wrapper function to console_menu
}
\examples{

console_file_menu("c/", ".txt")

console_file_menu("c/", ".txt", instruction_msg = "Please select a file to open")

console_file_menu(folder_path = "c/",
                 pattern = "xlsm",
                 instruction_msg = "Choose a excel file to open",
                 quit_key = "Q",
                 quit_message = "Type in Q and press return to quit"
                 return_number = TRUE
                 )
}
