% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_menu.R
\name{console_menu}
\alias{console_menu}
\title{Prints out a menu from a character vector and waits for user selection}
\usage{
console_menu(
  chr_vector,
 
    instruction_msg = "Please type in the console your choice from above and press enter: ",
  quit_key = "q",
  quit_message = paste0("To quit please type ", quit_key, " and press return"),
  return_number = TRUE
)
}
\arguments{
\item{chr_vector}{character vector}

\item{instruction_msg}{character.  Message shows under the menu selection}

\item{quit_key}{character.  Character value for the user to quit the menu}

\item{quit_message}{character.  Character value to explain how to quit the menu}

\item{logical.}{TRUE return number choice.  FALSE returns the index of chr_vector}
}
\value{
character
}
\description{
Prints out a menu from a character vector and waits for user selection
}
\examples{

console_menu(c("Eat", "sleep", "code", "repeat"))

console_menu(c("Eat", "sleep", "code", "repeat"), "What would you like to do?", "quit", "Type quit to exit")
}
