% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_menu.R
\name{print_menu}
\alias{print_menu}
\title{Prints out a character vector starting with a sequential number as a menu}
\usage{
print_menu(chr_vector)
}
\arguments{
\item{chr_vector}{character vector.}
}
\description{
Helper function for console_menu.
}
\examples{

\dontrun{
print_menu(c("Item One", "Item Two"))
}
}
